Fewer user-data parameters
Four `uX` parameters is overkill for a task.  We only need one.  I had
originally chosen four (or really, more than one) for the `flt_channel`
interface: I thought it would be nifty to have some of the parameters
provided by the channel constructor, and some by the sender.  But we can
mimic the same thing by sending a separate type into the channel
receiver, containing as many different fields as we need.  We now have a
bit of confusion over there being two separate things named `ud` in the
channel interface; it's probably worth renaming one of them.  But to
what...?