Thread-safe task groups
The task group functions are now (mostly) thread-safe.  (The main thing
that is still not thread-safe is marking that a group should be executed
after some other group has finished; this is thread-safe for the
“before” group, but not for the “after” group.)

In addition, the API has been cleaned up a bit; in particular, there are
now four explicit states that a task can be in, as described in the
flt_task(3) man page.