Type-specific flt_local accessors
Most of the flt_local functions now take in the type of the instances
being managed.  This is repetitive, but it means we have remove one
layer of pointer indirection in our implementation.  Since flt_locals
are used in a quite a few hot code paths, this actually has a noticeable
impact on the running time.

This means that instead of `new` and `free` callbacks, you provide
`init` and `done` callbacks, since the flt_local manager allocates a
single array of instances.  To eliminate false sharing, we pad each
instance to the size of a cache line (which we're currently hard-coding
as 64 bytes.)