Run one example at a time
The fleet-examples test program runs several different example
computations in several different fleet configurations.  This patch
updates it so that you can choose to run only one example, in one
configuration.  The usage is either:

    $ fleet-examples

to run everything, or

    $ fleet-examples [config] [example] [options]

to run a single example.  Each example takes in additional options to
choose how large of an input to run against, for instance.