Aligned flt_local instances
Each instance in an `flt_local` manager should be in a separate cache
line in memory.  This eliminates “false sharing”: if two instances are
in the same cache line, then the threads that own those instances will
“step on each other”, since they'll be bogged down sending cache updates
to each other.  By making sure each instance is in a separate cache
line, each thread is able to execute its code unimpeded.

We were already adding padding between each instance in the underlying
array, in an effort to push each successive element into a new cache
line.  However, we also need to align the array itself.  If we don't do
that, then each element, while spaced far enough apart, will fall on a
cache line boundary, guaranteeing that we'll have false sharing.