Consistent overflow in examples
All of our current examples calculate the sum of a sequence of integers.
This can easily overflow if the input size (i.e., the number of integers
to sum) is too large.  That said, we can make sure that the actual
calculation, and our simple multiplication (to get the expected answer),
overflow in a consistent way.  That way we can still verify that we're
getting the expected (overflowing) result.