Tasks can be singly- or doubly-linked
The internal task structure now contains fields for being either in a
doubly-linked list or a singly-linked list.  We need to have a
doubly-linked list for the ready queue of each execution context, since
we need to run from the front and steal from the back.  But for all of
the other cases where we need a list of tasks, the ordering of instances
in the list doesn't matter, and so a singly-linked list works just fine.
We can do this without using any additional space, and it's now
noticeably faster to allocate and free tasks via the context's unused
list.