Don't allocate tasks in batches
We no longer allocate tasks in batches, since we're going to assume the
rule of thumb that you don't allocate a huge number of task instances
all at once.  (We'll eventually provide higher-level APIs that let you
schedule a large number of tasks in a controlled way, with back
pressure.)

In addition to allocating each task instance singly, we now also make
sure that the new task instance is aligned to a cache line, to eliminate
one more possibility of false sharing.