Remove task count from task deques
We can shave some cycles by not maintaining the per-context task count
each time we schedule and run a task.  That means that we have to walk
through more of the deque each time one context steals from another, but
we're not optimizing for the migration case, so that shouldn't be a
problem.