Check for theft after every task
Before each context would execute up to 32 tasks in a row before
checking whether someone else was trying to steal from it.  I think
we'll want theft to be more responsive, so now we check after every
single task execution.  It looks like we can do this without a
performance penalty by not inlining the code that actually migrates the
results over to the burglar.